src/requirements.txt
numpy
matplotlib 
pillow
